Refactor range parser logic
