+++ lib/doc/image.js
@@ -1,5 +1,4 @@
-const colCache = require('../utils/col-cache');
-const Anchor = require('./anchor');
+const {parseRange} = require('./drawing-range');
 
 class Image {
   constructor(worksheet, model) {
@@ -36,22 +35,7 @@ class Image {
     this.imageId = imageId;
 
     if (type === 'image') {
-      if (typeof range === 'string') {
-        const decoded = colCache.decode(range);
-        this.range = {
-          tl: new Anchor(this.worksheet, {col: decoded.left, row: decoded.top}, -1),
-          br: new Anchor(this.worksheet, {col: decoded.right, row: decoded.bottom}, 0),
-          editAs: 'oneCell',
-        };
-      } else {
-        this.range = {
-          tl: new Anchor(this.worksheet, range.tl, 0),
-          br: range.br && new Anchor(this.worksheet, range.br, 0),
-          ext: range.ext,
-          editAs: range.editAs,
-          hyperlinks: hyperlinks || range.hyperlinks,
-        };
-      }
+      this.range = parseRange(range, hyperlinks, this.worksheet);
     }
   }
 }
